requirements.txt
prettytable==0.7.2
pyserial==3.3
